+++ Editor/AutoRunner/AutoRunnerWindowBase.cs
@@ -67,9 +67,9 @@ namespace SaintsField.Editor.AutoRunner
                 {
                     so = new SerializedObject(obj);
                 }
-#pragma warning disable CS0168 
+#pragma warning disable CS0168
                 catch (Exception e)
-#pragma warning restore CS0168 
+#pragma warning restore CS0168
                 {
 #if SAINTSFIELD_DEBUG
                     Debug.Log($"#AutoRunner# Skip {obj} as it's not a valid object: {e}");
@@ -104,9 +104,9 @@ namespace SaintsField.Editor.AutoRunner
                     {
                         so = new SerializedObject(comp);
                     }
-#pragma warning disable CS0168 
+#pragma warning disable CS0168
                     catch (Exception e)
-#pragma warning restore CS0168 
+#pragma warning restore CS0168
                     {
 #if SAINTSFIELD_DEBUG
                         Debug.Log($"#AutoRunner# Skip {comp} as it's not a valid object: {e}");
@@ -154,7 +154,7 @@ namespace SaintsField.Editor.AutoRunner
             // ReSharper disable once ConvertIfStatementToNullCoalescingAssignment
             if (_typeToDrawer == null)
             {
-                _typeToDrawer = SaintsPropertyDrawer.EnsureAndGetTypeToDrawers();
+                _typeToDrawer = SaintsPropertyDrawer.EnsureAndGetTypeToDrawers().attrToPropertyDrawers;
             }
 
             // if (SceneManager.GetActiveScene().isDirty)
@@ -247,9 +247,9 @@ namespace SaintsField.Editor.AutoRunner
                     {
                         so = new SerializedObject(extra);
                     }
-#pragma warning disable CS0168 
+#pragma warning disable CS0168
                     catch (Exception e)
-#pragma warning restore CS0168 
+#pragma warning restore CS0168
                     {
 #if SAINTSFIELD_DEBUG
                         Debug.Log($"#AutoRunner# Skip {extra} as it's not a valid object: {e}");