+++ Runtime/SepTitleAttribute.cs
@@ -12,7 +12,7 @@ namespace SaintsField
         public readonly string title;
         public readonly float gap;
 
-        public SepTitleAttribute(string title, EColor color, float gap = 2f, float height = 2f)
+        public SepTitleAttribute(string title, EColor color = EColor.Gray, float gap = 2f, float height = 2f)
         {
             this.title = title;
             this.color = color;
@@ -20,12 +20,8 @@ namespace SaintsField
             this.gap = gap;
         }
 
-        public SepTitleAttribute(EColor color, float gap = 2f, float height = 2f)
+        public SepTitleAttribute(EColor color = EColor.Gray, float gap = 2f, float height = 2f): this(null, color, gap, height)
         {
-            title = null;
-            this.color = color;
-            this.height = height;
-            this.gap = gap;
         }
     }
 }